requirements.txt
# MCP 기반 멀티 에이전트 LLM 고객지원 챗봇
python-dotenv>=1.0.0

# LLM / API
openai>=1.0.0

# 웹 앱 (필요 시 fastapi 등 추가)
# fastapi>=0.100.0
# uvicorn>=0.22.0

# RAG (필요 시 추가)
 langchain>=0.1.0
 langchain-community>=0.2.0
 langchain-openai>=0.1.0
 chromadb>=0.4.0
 sentence-transformers>=2.2.0
 faiss-cpu
# 거버넌스 / PII (필요 시 추가)
 presidio-analyzer>=2.2.0
 presidio-anonymizer>=2.2.0
